Address PR review: move ForecasterModules out of forecasters/ into modules/
BaseForecasterModule, DeterministicForecasterModule and
ProbabilisticForecasterModule (Lightning wrappers around a Forecaster)
were mixed in with the Forecaster classes themselves under
neural_lam/models/forecasters/. Move them into their own
neural_lam/models/modules/ package (base.py, deterministic.py,
probabilistic.py) so the two concerns: what a forecaster is vs how
it's trained/evaluated by Lightning live in separate directories.
Pure move: internal imports and neural_lam/models/__init__.py updated
accordingly, no behavioural change.
