+++ neural_lam/models/__init__.py
@@ -3,13 +3,13 @@
 # Local
 from .forecasters.autoregressive import ARForecaster
 from .forecasters.base import Forecaster
-from .forecasters.base_module import BaseForecasterModule
-from .forecasters.deterministic_module import DeterministicForecasterModule
 from .forecasters.probabilistic import (
     ProbabilisticARForecaster,
     ProbabilisticForecaster,
 )
-from .forecasters.probabilistic_module import ProbabilisticForecasterModule
+from .modules.base import BaseForecasterModule
+from .modules.deterministic import DeterministicForecasterModule
+from .modules.probabilistic import ProbabilisticForecasterModule
 from .step_predictors.base import StepPredictor
 from .step_predictors.graph.base import BaseGraphModel
 from .step_predictors.graph.graph_lam import GraphLAM
