+++ neural_lam/models/modules/__init__.py
@@ -1,0 +1,8 @@
+"""
+Lightning modules wrapping forecasters for training and evaluation.
+"""
+
+# Local
+from .base import BaseForecasterModule
+from .deterministic import DeterministicForecasterModule
+from .probabilistic import ProbabilisticForecasterModule
