+++ tests/test_train_model_warnings.py
@@ -77,7 +77,7 @@ def test_create_gif_forwarded_to_forecaster_module():
         patch("neural_lam.train_model.MODELS", {"graph_lam": MagicMock()}),
         patch("neural_lam.train_model.ARForecaster"),
         patch(
-            "neural_lam.models.forecasters.deterministic_module."
+            "neural_lam.models.modules.deterministic."
             "DeterministicForecasterModule.__init__",
             capture_init,
         ),
